# Pre-compute powers of two to speed up cardinality estimation