When benchmarking normal HLL cardinality estimation, I noticed a lot of time is spent doing `math.Pow`. Since that's only done on `uint8` values, it makes sense to compute it in advance and use a lookup table. It uses two kilobytes of memory, which I think is worth the ~5x speedup.