@mynameisfiber 2018-10-09
For posterity, below are the benchmarks before and after this change. In particular, there is a 5x speed up on the `BenchmarkCardinalityNormal-4` benchmark. Well done @kokes !

Before:
```
gohll:(master)$ go test -bench=.
BenchmarkAddSparse-4             1000000          1767 ns/op          93 B/op          3 allocs/op
BenchmarkAddNormal-4            10000000           226 ns/op          16 B/op          2 allocs/op
BenchmarkAddSparseFNV1A-4        1000000          1564 ns/op         101 B/op          4 allocs/op
BenchmarkAddNormalFNV1A-4        5000000           244 ns/op          24 B/op          3 allocs/op
BenchmarkCardinalityNormal-4         200       5953826 ns/op           0 B/op          0 allocs/op
BenchmarkCardinalitySparse-4        2000        786735 ns/op          73 B/op          0 allocs/op
PASS
```

After:
```
gohll:(kokes-patch-2)$ go test -bench=.
BenchmarkAddSparse-4             1000000          1770 ns/op          93 B/op          3 allocs/op
BenchmarkAddNormal-4            10000000           229 ns/op          16 B/op          2 allocs/op
BenchmarkAddSparseFNV1A-4        1000000          1580 ns/op         101 B/op          4 allocs/op
BenchmarkAddNormalFNV1A-4        5000000           244 ns/op          24 B/op          3 allocs/op
BenchmarkCardinalityNormal-4        1000       1184037 ns/op           0 B/op          0 allocs/op
BenchmarkCardinalitySparse-4        2000        791731 ns/op          73 B/op          0 allocs/op
PASS
```