@kokes 2018-10-11
It just occurred to me that having 256 powers is excessive, we won't ever need more than 64 (or even less, considering that a few bits determine the bucket). That is unless a 256-bit hash function is planned in order to support cardinalities of 2**200 and beyond :-)